requirements.txt
anthropic
openai
cloudinary
typer[all]
pydantic>=2.0
pydantic-settings
python-dotenv
rich
watchdog
python-slugify
pyyaml
httpx
